requirements.txt
django==1.8
django-mptt==0.8.1
django-image-cropping==0.8.3
easy-thumbnails==2.1
pytils==0.2.3
djangorestframework==3.3.3
django-filter==0.13.0
markdown==2.6.6
